OBPIH-7291 Improve logic and remove testing functions
