+++ src/js/components/cycleCount/cycleCount.scss
@@ -242,11 +242,6 @@ input.hide-arrows[type=number] {
 
 
 .modal-content {
-  //position: absolute;
-  //top: 50% !important; /* Center vertically */
-  //left: 50% !important; /* Center horizontally */
-  //transform: translate(-60%, -100%) !important; /* Center both axes */
-  width: auto !important;
   min-width: 1000px;
 
   .assign-count-modal-header {
